requirements.txt
python-dotenv==1.1.0
aiogram==3.20.0.post0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
googletrans==4.0.0-rc1
deep-translator==1.11.4
openpyxl==3.1.2
Flask>=2.0
playwright
lxml